create networks, routers, and fip, and boot server and ping test it.
the following class is defined in neutron-netcreate_nova-boot-fip-ping.py:

  NeutronBootFipPingPlugin

    with the class method :
      create_network_nova_boot_ping()

    and a modified version of
      _create_router() (used by above method)

copied from NeutronScenario() class.  We needed a way to
designate a specific external network in the case where
openstack may have multiple external networks defined.

The json example file neutron-plugin-fip-ping.json
illustrates how to use this plugin and can be called
with:

rally --plugin-paths ./ task start ./neutron-plugin-fip-ping.json

args needs to include "ext_net" as a dict with key "id" as in the
example.