fix an issue with data gathered from redis -- kambiz
This script is meant to work with collectd's "exec" plugin.
The data collection contains several dirty variables in that
the values stored contain "\n\r".  As such, the used_memory
PUTVAL which contails a : causes the : character to move to
the front of the string.  This is bash behavior and can be
fixed by stripping the extra  from the strings.

Change-Id: I9c5e08c1ac62ed176ead4740146398f49d0f7b1a

fix collectd-redis.sh

minor fix from previous patch.

Change-Id: Icc35644a3de5895372cc544c2984ed0d8cdba02c

fix

Change-Id: I5a3ee6266e7c63ada5718868dea2394827b1f85a