install graphite and grafana as docker containers
This adds additional playbooks to deploy:
   - carbon-cache
   - graphite-web
   - grafana

as containers.  The data should map to the host running dockerd
for data persistence across reboots.
Ensure:
   - persistent data if containers are restarted
   - systemd scripts to ensure containers startup
   - new ports defined in the group_vars/all.yml
   - fixed mismatch between templates and variables
   - minor space issues to keep folks happy
   - Use a help script for docker ps to avoid weirdness with
     '{{.Names}}' in the command line when using the command plugin.

These have been tested on RHEL7 hosts, as well as undercloud
deployed via RDO quickstart.  The following variables are
of note in the group_vars/all.yml and how to use them:

Location on the target docker host where whisper data is kept:

     persistent_carbon_data_path: /data/carbon/whisper

Location on the target docker host where grafana runs to ensure
persistence of grafana related data across reboots (e.g. dashboards, etc):

     persistent_grafana_data_path: /data/grafana

This port is mapped from the docker host to the standard carbon-cache port
on the launched container (basically makes it easy to ensure ports don't
clash in case there is anothre service listening here):

     docker_carbon_cache_port: 2003

This is the web port on the docker host that is proxied back into the
container (useful when there are a number of other web services running,
e.g. if deploying on an undercloud host):

     docker_graphite_port: 8888

Same with the following.  It's not likely there will be clash with this
port but just in case you are running an instance of grafana, this
allows mapping an alternate port to the docker container:

     docker_grafana_port: 3000

This is the published docker image for carbon-cache.  This is based on
the centos7 image using python-carbon rpms.  This is published into
hub.docker.com.  See:  https://hub.docker.com/r/kambiz/carbon-cache/

     carbon_cache_docker_image: kambiz/carbon-cache:0.9.15

This is the published docker image for graphite-web.  This is based on
the centos7 image using httpd and graphite-web rpms.  This is published
into hub.docker.com.  See:  https://hub.docker.com/r/kambiz/graphite-web/

     graphite_web_docker_image: kambiz/graphite-web:0.9.15

This is the official grafana project docker image.  The image will never
change since we are using 2.6.0, rather than "latest".  See the images
here:  https://hub.docker.com/r/grafana/grafana/

     grafana_docker_image: grafana/grafana:2.6.0

Change-Id: Icf63d0b044cded56cbecc7d306847dedd84bec87