Replace dummy 1.1.1.1 values for graphite_host, grafana_host
This adds a sanity check that will fail if the user has not filled out
the grafana_host and graphite_host values in group_vars/all.yml

It will cause the grafana playbook execution to fail with a debug warning
which is cleaner than having it execute all the way through only to fail later.

We would ship these entries defined but empty as a result.

Patchset #2: correct typo in commit title
Patchset #3: move to top of playbook so if it fails it will be right away.
Patchset #4: README.MD update: specify that these values need to be filled in

Change-Id: I333798398ee0a84d1af51b9344b7c4798aad2043