Use -tt in gen_hostfile.sh to force CI compatibility
Using ssh with -t forces the allocation of a pseudo terminal
server side it does not create a terminal if there is no client
side tty. The -tt option resolves this issue by forcing the
creation of a pseudo terminal on both ends if required. This
is important for CI programs like jenkins where the Browbeat
install scripts may be executed without tty.

Change-Id: Iae892f3462ed6bbdab00f782e8bec1cf4ad1431f