update README for docker svc installs
Specifically when the graphite, carbon-cache and grafana services
are installed on undercloud host, the user should know to use the
docker playbooks.  This is to avoid package version requirement
issues.  This is a change to the README.rst to make it more clear
to the end user when to use which playbook.

README can be viewed here:

  http://rst.ninjs.org/?n=56d6bb46eadab514752335f7e11e3881&theme=basic

Change-Id: I929dd840b22a0342a5044aadb4e5141af936d1fa