Disable known hosts for generate_tripleo_hostfile.sh
This pull is to solve an issue with the Browbeat CI job
where we may be assigned arbitrary machines from a pool
with keys that may or may not remain the same.

I'll note that the ssh config files generated by this script
already setup the undercloud ssh shortcut to skip known hosts
verification, so this does not introduce a MITM vulnerability.

That being said if Browbeat where to target production use at
any point in the future this change, as well as the ssh config
file generated would need to be corrected for secure default
behavior.

Change-Id: Ice8a7e3c3b99a8564321b938e631122ede4942b0