Added Ceilometer metadata
This change is now tested and shown to be working.

This commit adds Ceilometer metadata collection
to the gather playbook. The variables chosen
for collection where selected by going through
the config file and grabbing what looked relevent
please provide feedback as to what is and is not
relevent for collection.

On another note, this file is somewhat long, thoughts
on moving to a strcuture for the gather playbooks
that would involve looping over a list of variable names
rather than individual tasks for each var collected?

Change-Id: Idf3f75f152947f201ea75966e4303a0787555d9e