Add cinder metadata
This commit adds collection for cinder metadata.

I am working on a refactor of Browbeat metadata in general
but after spending today working on it, it's not going to be
done very soon. Since this was already on my hard drive and works
I figured it was better to get it out where people can use it even
if an improved version should be coming soon.

Change-Id: Ib11aa1056e89e1e42de7e514d8ac304cb8d4b42b