Fix collectd install
The final check in the collectd install playbook
checks that collectd is shutdown on the overcloud
compute node, this ignores the fact that collectd is
probably not installed on the compute node, so the check
fails to turn off the nonexistant instance of collectd
and fails the playbook.

The commit fixes the issue, it might be better to
check if collectd is installed and then try and turn it
off if it is. But I don't see why turning it off should
fail so long as we have proper permissions.

Change-Id: I45f1efca2c257987ab46f87a1dcfb19ae0004e5d