Fix sudo requirement for localhost
All roles in this playbook previously had
become: true, which requires sudo permissions
the localhost task, running the metadata.py file
do not require root. When being run from the undercloud
this is not major issue as the undercloud and locahost
are the same and passwordless ssh exists on both.

But when running Ansible/Browbeat remotely this will
require root permissions unnessicarily.

Change-Id: I71e97348dcce25ebb9ec2aa9a00fce516315177c