Correct template default values for Rally plugins
These Rally plugins are designed to take num_networks as
an optional argument, where num_networks=1 if not otherwise
defined.

https://github.com/openstack/browbeat/blob/master/rally/rally-plugins/netcreate-boot/netcreate_boot.py#L29

But when Rally takes the Browbeat config values and attempts to template
the yml files in the plugin folders num_networks is not set to a default
value in those template files, meaning Rally fails to generate the templates
and to ever reach the plugin which wouldn't care anyways.

This change adds a default value for num_networks to the plugin template files
such that when Rally templates off of these files it will run successfully regardless
of if num_networks is defined.

Note that from the perspective of the Rally plugin it will now always recieve a
value for num_networks even if that value is 1, the same thing it would be if it
had not recieved a value at all.

Change-Id: I4313aaed963be1689be91bbd180a046e70f987b3