Updated method to index hosts for Grafana Dashboard creation
Updated the hosts_to_dictionary function to check to see if the last
digit of the overcloud host name is a digit.  If so, it uses a regex
to isolate the all of the digits at the end of the hostname, and proceeds
to use that as the dictionary index.

This change allows for alternate overcloud host naming schemes that do not
have a hyphen preceding the server number.

Change-Id: Ief10b2100b5d7c55a940ab4186ce2e31c109420c
Closes-Bug: #1612787