Fix rally_version in CI
The rally version was added as a variable in the
install group vars file, but CI maintains it's own
version of this file that was not updated, therefore
the rally_version change was breaking to all CI builds.

It might be better for CI to move to overwriting specific
variables if possible.

Change-Id: Ib4b789cc2cdbe409c5a866444fe144328f086e8a