Fix Metadata Generation
With the current model of dumping ansible facts with the template and
regexing it to make it JSON readable, metadata generation into files
is failing in some cases. It is better to have the template handle
creation of a serializable JSON object and that is what this commit does.
Also some unnecessary keys in lib/Metadata.py used to parse the JSON have
been removed since all of that is now handled by the template.

Change-Id: I7d4a4f9fde0af8a52aa72374614c20f4b8abc421
Co-Authored-By: Alex Krzos <akrzos@redhat.com>