Handle case when metadata files are configured but not present and also case where metadata file list is empty
This commit specifically fixes the code to handle the case where metadata files
are configured in the browbeat configuration file but are not present in the
metadata directory. Previoulsy since the combine_metadata method return boolean
value false in such cases and the index_result method treated result as a
dcitionary, traceback was seen. With this commit, combine_metadata stops
returning false and instead exits the run if an intended metadata file is
absent.

+ Changing how we return result so that the code doesn't err out even when
  metadata file list is empty in the confing file
+ Checking in browbeat.py too
Change-Id: I52712beaa2dec6209394a2f3ef605d5a9a13f5cb