Make Elasticsearch connector more fault tolerant
Previously, if elasticsearch was enabled and sending data to it failed,
browbeat would end with an ungly traceback. This commit introduces a try,except
statment and also adds a retry for sending to elastic. Failing 2 retries, the
result JSON is written to the results directory(for example,
results/20160920-215128-browbeat-authentic-keystone-24-iteration-0-elastic.json).
The user can later curl this to the Elasticsearch instance.

+ Converting browbeat_uuid and timestamp to str, since thos objects are
  not JSON serializaable and can't be written to a JSON
+ Retry and sleep interval are fixed for now, can make it configurable
  if need arises

Change-Id: I2f5a173490bf93d44de6a9ef03c0aeb1b6a8eaac