Fix generate_tripleo_hostfile.sh
Since there is a error message now being sent to stderr the fix is to
remove that message by sending it to /dev/null. The OpenStack client
from OOO needs to be updated to properly fix this. However, this fix
should not impact other releases without this issue.

Fixes bug 1626324

Change-Id: I2f88642f34f501e50f0e7f2d17ae229d311e0dd3