Make metadata robust to missing services
So currently if some optional openstack service is not installed
metadata will fail when it tires to work on its config file.

This commit adds in a check for the existance of each service config
before running the gathering tasks.

What might be a useful addition to this commit would be some logic
for determining which services should never be missing and alerting
the user.

Change-Id: Ie360f062646cb5431de9ed21bd0dd193b84ddcda