Add ovn services to collectd role.
This patch adds three new services that are part of OVN to the
collectd role.  OVN is a sub-project of OVS, so I added them to
the existing openvswitch section.  If we split them, it's not
clear what to do with ovsdb-server, which is used by ovs-vswitchd as
well as all of the OVN services.

I'm also assuming this change won't cause any problems if OVN is not
in use.  Even if OVN is in use, ovn-northd and ovn-controller will
certainly exist in the environment, but ovn-controller-vtep may not.

Change-Id: I6f5c1584c700a450c4090e831fb51fd650e5bba0
Signed-off-by: Russell Bryant <rbryant@redhat.com>