rally/subnet-router-create: Fix multiple networks.
This code had two cases, one for when there was only 1 network and
another case for when more than 1 network was used.  Simplify the code
so that only one code path is needed.  This also fixes an off-by-one
error in the multiple networks code.  It was actually only doing N-1
networks instead of N, which could explain why an alternate code path
was added for a single network, as this code would not have done
anything when num_networks == 1.

Tweak some spacing in passing to be pep8 compliant.

Change-Id: I27cac72f1e20774e8f097fe5e24b20fa6b76970e
Signed-off-by: Russell Bryant <rbryant@redhat.com>