Collectd plugin: Gnocchi status
Initial commit to add a plugin to collectd for obtaining and storing
Gnocchi status metrics.  This plug obtains the two metrics provided
by Gnocchi's REST API for "the number of metrics having measures to be
processed" and "the number of measures to process".

This also addes an associated graph in the General OpenStack System
Performance Dashboard and the Controller specific dashboard.

* no longer requires authentication details in the collectd config
* removing the redis exec plugin
* configurable interval for the python plugin now

Change-Id: Idfc5c719e9b0fbaad5db0632a3e5293bf03d7ab6