Initial cleanup of Browbeat documentation
All this commit does is move the contents from the README.rst file
into the various subfiles within the tox documentation source folder
so that things render cleanly and with a proper table of contents.

I've resisted the urge to start copy editing, adjusting content,
fixing spacing, titles, etc. All changes should be a simple copy/pasta
from the README to the various subfiles.

I did add some minor content to the README.rst so that it wasn't
just a blank file for those who land there via GitHub or a git
clone.

Change-Id: I892e4e328a2715ddcbafedd97d1328d1f709f312