Store all Rally results under common "rally" directory
This is going to help us easily identify the workload for post processing
of JSONs and remove the need for complex regexes and logic in lib/Tools.py for
post-processing. Since we have the flexibility to pass *args now, I am also deleting an
unused method which is no longer needed.

Change-Id: I54e601dde25bba6e5239f4ee1d3a61893cdb50a1