Fix source for CI vars file
The CI vars file is on the jenkins slave, not on the virthost where we
where trying to copy the data from. The code that was checked in has never
worked, what happened is that with a previous iteration of the commit we
created /tmp/install_vars.yml and since the jenkins machine never gets rebooted
it just stayed there. So in the future when we tried to fetch the file from the
wrong machine Ansible failed the fetch silently and we loaded the old file.

Change-Id: I8c20bdc8dc138a383e7a101cd465b48d44fa99a7