Use only yum Ansible in CI setup
So for months now Ansible sourced from yum and Ansible sourced
from pip could exist on the same machine and function just fine.
But as of a few days ago in the Newton and now rhos-10 CI pipeline
the dependency loading semantics have changed and this is no longer
possible.

This commit switches to relying on the package manager version.

Change-Id: I21737c6f6900f3ed4011fe7de91fc34a1d984fc3