Store link for Jenkins build logs in CI version metadata
So this is a very easy addition on this end, the problem is that right
now logs are only collected for two weeks before being removed. Not to
mention this link won't work until the build is done. So there are a
couple of ways someone could click this link and get a 404.

I'll talk to weshay about getting longer term log storage.

Change-Id: I01fa5ea4b2560d918b82f9d12990d63fbe379f3e