Adjust CI benchmark run to not knock over the endpoint
https://thirdparty-logs.rdoproject.org/jenkins-browbeat-quickstart-ocata-baremetal-no_vlan-46/undercloud/home/stack/browbeat/results/20170518-220119/all-rally-run-0.html.gz

So if you look at these test results you can see that the
12 concurrency doesn't even run because we knock over the
endpoint. We only have two quad core computes for this and
they always get knocked over. Another problem is that Browbeat
only checks if the rally task generates a file, so if the endpoint
is down this isn't described as a test failure, but if the endpoint
returns an error (arguably better than being down) Browbeat will
fail the test. Overall I don't like the randomness of this.

Change-Id: I27f80c8f9acc77c74a40a715e28af54ec3354456