Bulk indexing of all elastic docs
This creates a general purpose Elastic object cache in the Browbeat
Elastic class. Every call to insert pushes into a cache which is then
emptied at the end of the workload run, when the cache max
size value is full or when more than a given amount of time has passed.
By default 1000 entires or 10 minutes. With a final flush being called
on the destruction of the Elasticsearch object. (even gets called on
sigterm so in theory you can ctrl-c and save your results)
It is emptied using a 4 parallel workers and the bulk api.
So it should be efficient at inserting large amounts of data.

If the upload of any one of the objects fails the rest will not be
inserted and the entire cache, included previously successful inserts
will be written to the disk. Possible improvement is to have the helper
return a list of pass/failed inserts and dump only those objects.

Change-Id: I0fceb5888c6f7d3167320593177d9cdc72504878