Fix some errors that slipped by the linters
So up until recently the 'linters' CI run ran both
the pep8 and the Ansible linters, after some review
I can only conclude that it actually was not supposed
to do that and I should have seperate Python and Ansible
linters jobs.

Anyways because of this the pep linters where out for a
few days an we got exactly two minor pep errors into our
repo.

Change-Id: I66f82021c71f89a775e05c411ee1af2dba3a1c7b