Refactor CI playbook into sub playbooks
So originally this commit was for ceph compatible gather
but I was never super confident that my solution was good
on the other hand I made a useful refactor while working on
that which I want to keep and commit.

This breaks the previously monolithic CI playbook into smaller
playbooks that can be composed more easily. Hopefully useful
as we move to seperating integration and performance testing.

Change-Id: I9fe48a65861caacaf8092d52d837801eec246fb0