Integrated Browbeat-ML to classify tests
This adds a perf classification task to run after each
performance CI run, if any of the results are F (failing)
the test returns 1 and the Jenkins build will fail.

For reference think F = 1, D = 2, ... 5 = A.
Think of it as a simple curve grading scheme for performance
if a test gets a 5 for a test it's the best performance
of the 'class' for the past couple of months.

Change-Id: Id4f1ae36fd33ecc9b0638ff1b8b46497b3c9e34b