Prep CI for Pike
So this detects if a deployment was containerized and then
stores that in the metadata. This also disables results in httpd
since that seems to cause problems for Pike deployments at the moment

The final change is some sort of permissions problem with running ping
during the browbeat install. I'd like to find a better solution than
just slapping sudo on the front though.

Change-Id: I57c0006bc9abe861782ebdfb918b4151bba67eae