Set Browbeat-ML to upload summarized data at the end of a test run
Sets the upload option to for Browbeat-ML to true, this uploads
aggregate info to a database we use to quickly access long term data
rather than spending hours crawling Elastic and performing a summary
or crawling Graphite only to lose the data at the end of the retention
period.

Essentially if it's pretty simple to gather but time consuming to do so
this commit will go ahead and drop it in a sql db right after we run the
test for us to poke at later.

Change-Id: I484e8d58e533b5cbebc8403d4ef26a4ab666381a