Add options to support varied stack names
This change adds the ability to generate an inventory file and
ssh config when using stack names other than "overcloud".

The script still defaults to "overcloud" if the options are
not specified.

Also now sets the variable $tripleo_ip_address to "localhost"
if the localhost option is specified. Prior to this change,
specifying -l would leave the variable empty which would fail
the first if statement and only give you the script usage.

Change-Id: Ia7fc1bc552dcf7a8ece1ad5b79153e0d87a39537