Fix non multiple of four indentation in Tools and Yoda
So turns out that E111 is the multiple of four indentation check
it's my fault that ended up in the tox.ini last time I was messing
with it and it seems to have come back to bite only me, as Yoda and
Tools functions, also for Yoda are the only victims of bad indentation.

After a strong recomendation by a friend I'm giving autopep8 a try.

Change-Id: Ia067bc60fad1200f577624aac2e979dfc26ec05d