Update stackrc parser for Pike changes
The Pike stream has a new stanza in the stackrc
that adds the cloud name to PS1 if the variable
CLOUDPROMT_ENABLED is set, the if statement screws
up the YODA stackrc parser.

This just ignores that as we don't care about
anything that's not the auth_url, username
and password.

Change-Id: Ia6fc2f9aed151d5a9d08ec00a25ba161bb93d624