don't spend hours testing things with WIP in the title
So this is very easy, the problem is determining if the user
actually intended to mark their commit as a work in progress
for no testing, or if the user has WIP in their commit title
for some other reason. To help with that WIP or wip has to
be at the start of the commit message and have a trailing space.
Maybe : should also be accepted as a trailing char.

Change-Id: Ic44114dbf50abbe2a9f135a41257208fe7feddf8