Resolve upstream breakage from oooq changes
So TripleO is moving to not building undercloud images. In doing so 
they used the obvious name for a dictionary of images just like
we did. This causes issues since the config vars take precedence over 
our own group_vars. Replace with more discriptive var name.

Also changed are the meadata gathering functions to use the overcloud
image instead of the now non existant undercloud one.

Finally collectd openstack has to handle the inclusion of a different
version of collectd by default now. 

Change-Id: Ie29c1c2070ffe88e015a1a04bdb4834cf14f1e5c