Add ryslog stop statement at parsing end
For some osp versions there are more statements after the stop
statement in 05-outputs.conf. Since config files are loaded in
lexical order and we don'totally blow away the existing config
the only way to be 100% sure that stop is the last thing
loaded is to insert it at the end of the main config.

This causes a bug in the adiscon repo. But the official CentOS versions
of all the packages we need are finally available. So we switch to those

Change-Id: I2109bde6572cc8ccf7d633ed1a795bf8449d1a89