Add mechanism to gracefully exit on SIGINT
This commit adds functionality for a user to exit out of a long running
browbeat configuration by using SIGINT. On catchinga  SIGINT, browbeat
compeltes the current workload and then exits.

To keep the code lightweight, browbeat termination on SIGINT happens
only at a workload level. The current running workload (in case of
rally, the scenarios and concurrencies associated with the workload) is
finished.

Change-Id: Iaa1f10233dabd053293327f42f8bd1320f1af95d