Clean up glance before moving the workload guest images to glance
When the browbeat playbook is run with workloads enabled, new images will be created for the workloads and moved to glance. When the playbook is rerun, it fails because the image already exists in the same name in glance. So we need to clean up the old workload images in glance(if they exist) so the new ones are moved to glance without any errors.

Change-Id: I766143c4f329a600405c705d29252fee7d22bd8c