Add virtualization metadata and move kernel to software
This commit adds two new fields in hardware metadata to
identify if the node is baremetal/vm. This will help when running
browbeat using infrared in vms or when the control plane is running
on RHEV/oVirt.

This commit also fixes the long known issue of having kernel details
in hardware metadata.

Change-Id: Ie9dd9d375125ebeae0287c624d345b272e19e459