Switch to ranges for dependencies to avoid conflicts in consumers
tripleo-quickstart lists browbeat as a dependency and too strict
requirements can cause conflicts. Avoid case where ansible version
used by tripleo was downgraded to a buggy old version mentioned
by browbeat.

To be clear: if needed, it would be expected to add version capping
here. If we know that ansible 2.9 will break the code it would be
wise to add "<2.9" to it. Adding testing with newer versions would
be subject of another change.

If you add a capping please always add a link to the bug that made
it necessary, it helps a lot later.

Change-Id: I163a7088d687dbaf0a04f4b9c2f706652a92c7c4