Fix _associate_floating_ip() conflict
Both NeutronScenario and NovaScenario classes contains a method called
_associate_floating_ip() with different signatures.

The CreateNetworkNovaBootPing test class inherits from NeutronScenario
and VMScenario (which inherits from NovaScenario).

When running the test, I was getting the following error:

TypeError: _associate_floating_ip() got an unexpected keyword argument
'fixed_address'

And turns out the _associate_floating_ip() from the NeutronScenario
class was taking precedence and it doesn't contains the "fixed_address"
parameter.

This patch is inverting the inheritance order in the
CreateNetworkNovaBootPing test class so that the
_associate_floating_ip() from NovaScenario is the one being invoked.

Change-Id: I139664e2cc8ecc76e80a9eef286d733504734016
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>