ping scenario with security groups
This scenario enables to boot VMs with security groups. User can specify
number of security groups and security group rules to be created per
iteration.

As we know, only members and not rules of remote security group are used
in security group implementation, this scenario creates a remote
security group for each user requested security group but with no rules.

Additionally it creates a security group with ssh and icmp rule
for each iteration.

After the VM is booted, this scenrios tests pinging VMs floating ip.

Change-Id: I72e3dc826a1904894e39a2e756d1800cb7db20e1