trunk subport network simulation
Existing rally tests are only creating subports but not testing
if the real network connection established to subports or not.
In this patch, we create a vlan interface inside the VM for
each subport and ping this vlan interface.

Test will create 2 VMs and trunk subports inside these VMs.
Then it will ping the subport floating ip from the other VM.

Later it swaps the subports floating ips and retry the ping
test after swapping. This is to test movement of floating ip
across the VMs.

Though VM will have multiple trunk subports, we are testing
connection (i.e ping test) for the first subport.

Co-authored-by: Sanjay Chari <schari@redhat.com>
Change-Id: Id317d9c50ad914d1bef370488a9abc967081f5b6