Tag functions for dynamic workloads servers
This patch introduces functions for booting servers with tags, which can be used
to efficiently list required servers in various scenarios. It also adds a function
to retrieve servers based on a specific tag. The nova_custom.py and neutron_custom.py
files are temporary, and will be moved to a common dynamic_utils.py file in subsequent
patches.

Change-Id: I4f8d432ff66fa17cea1316169cb71f586727a061