Change in router for VM dynamic workloads
This patch introduces the following changes.
1. A router is created for each iteration for VM dynamic workloads,
and this router is used for all VM dynamic workloads for that iteration.
2. The router, network and subnet from Rally context are used every 5
iterations for VM dynamic workloads. The network and subnet from Rally
context are used instead of creating a new network and subnet for that
iteration, as the router from Rally context will be deleted after the
network and subnet created in that iteration, and this leads to conflicts
in cleanup.

Change-Id: Idbe1b7210042a81ac5ac9371974ad5fe86b5d2ef